pyproject.toml
[tool.poetry]
name = "cs238-final-project"
version = "0.1.0"
description = ""
authors = ["Danny Kolano <dkolano2@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7, <3.10"
"agi.stk12" = {path = "bin/agi.stk12-12.2.0-py3-none-any.whl"}
numpy = "^1.21.4"
ipykernel = "^6.6.0"
tensorflow = "^2.7.0"
tf-agents = "^0.11.0"
matplotlib = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
